Move literal tests into separate folder
